+++ packages/svelte/src/internal/client/dom/elements/attributes.js
@@ -20,7 +20,7 @@ import { clsx } from '../../../shared/attributes.js';
 import { set_class } from './class.js';
 import { set_style } from './style.js';
 import { ATTACHMENT_KEY, NAMESPACE_HTML } from '../../../../constants.js';
-import { block, branch, destroy_effect } from '../../reactivity/effects.js';
+import { block, branch, destroy_effect, effect } from '../../reactivity/effects.js';
 import { derived } from '../../reactivity/deriveds.js';
 import { init_select, select_option } from './bindings/select.js';
 
@@ -491,7 +491,7 @@ export function attribute_effect(
 		var current = set_attributes(element, prev, next, css_hash, skip_warning);
 
 		if (inited && is_select && 'value' in next) {
-			select_option(/** @type {HTMLSelectElement} */ (element), next.value, false);
+			select_option(/** @type {HTMLSelectElement} */ (element), next.value);
 		}
 
 		for (let symbol of Object.getOwnPropertySymbols(effects)) {
@@ -513,10 +513,12 @@ export function attribute_effect(
 	});
 
 	if (is_select) {
-		init_select(
-			/** @type {HTMLSelectElement} */ (element),
-			() => /** @type {Record<string | symbol, any>} */ (prev).value
-		);
+		var select = /** @type {HTMLSelectElement} */ (element);
+
+		effect(() => {
+			select_option(select, /** @type {Record<string | symbol, any>} */ (prev).value, true);
+			init_select(select);
+		});
 	}
 
 	inited = true;